% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes.R
\name{adjust_x_continuous}
\alias{adjust_x_continuous}
\title{Helper for adjusting a continuous x axis}
\usage{
adjust_x_continuous(end, end_nudge, mids = NULL, labels = waiver())
}
\description{
Helper for adjusting a continuous x axis
}
