% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{bars_in_bars}
\alias{bars_in_bars}
\title{A bars within bars chart}
\usage{
bars_in_bars(
  data,
  x_outer,
  x_inner,
  y,
  outer_fill = "#002856",
  outer_label_pos = 35,
  outer_width = 0.9,
  inner_fill = "#009AD7",
  inner_label_pos = 15,
  inner_width = 0.75,
  y_wrap_width = 20,
  text_size = 12,
  bar_width = 0.75,
  x_end = NULL,
  x_end_nudge = 1,
  x_mids = NULL
)
}
\description{
A bars within bars chart
}
