% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_dotplot.R
\name{change_dotplot}
\alias{change_dotplot}
\title{A dotplot displaying change over time}
\usage{
change_dotplot(
  data,
  y,
  from,
  to,
  change_col = TRUE,
  sort_by = change,
  change_col_x = 8,
  dot_size = 3,
  label_spread = 3.5,
  x_lims = c(0, 110),
  x_breaks = c(0, 50, 100)
)
}
\arguments{
\item{dot_size}{The size of dots}

\item{label_spread}{How far away from the dots you would like
corresponding data labels to sit}

\item{sort}{The name of the variable to sort y-axis labels by. Can be any
variable passed to \code{to} or \code{from}, the name \code{change} (if
\code{change_col = TRUE}), or \code{NULL}, in which case it sorts by the default
for \code{y}'s data type.

Need to include an error message for when sort = change and
change_col = FALSE}
}
\description{
This differs from other dotplots in that dots are connected by an arrow
conveying direction from "from" to "to"
}
