% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{normal_bars}
\alias{normal_bars}
\title{A generic template for a layer of "bars + data labels + neat y-axis"}
\usage{
normal_bars(
  x,
  y,
  sort = TRUE,
  bar_width = 1,
  bar_fill = "#002856",
  label_color = "#000000",
  label_weight = "plain",
  label_pos = 10,
  y_wrap_width = 20,
  label_size = 12
)
}
\description{
\code{label_pos} represents a single unit movement of the label along the
x-axis. Bear in mind that x-axes with small units (i.e. from 0 to 1) will
need only small nudges, while larger units (i.e. 0 to 100) will need larger
ones.
}
