% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{one_bars}
\alias{one_bars}
\title{Create a One-Variable Bar Chart}
\usage{
one_bars(
  data,
  var,
  sort = TRUE,
  complete = FALSE,
  exclude_y_axis = FALSE,
  question_subtitle = TRUE,
  label_pos = 10,
  y_wrap_width = 30,
  x_end = 100,
  subtitle_wrap = 80
)
}
\arguments{
\item{data}{A data frame}

\item{var}{The variable from the dataset to be plotted as bars.}

\item{sort}{Logical, whether the bars should be sorted. Default is \code{TRUE}.}

\item{complete}{Logical, whether to include all factor levels, even those not present in the data. Default is \code{FALSE}.}

\item{exclude_y_axis}{Logical, whether to exclude the y-axis text. Default is \code{FALSE}.}

\item{question_subtitle}{Logical, whether to include the variable's label as a subtitle. Default is \code{TRUE}.}

\item{label_pos}{Numeric, position for the labels on the bars. Default is \code{10}.}

\item{y_wrap_width}{Numeric, width for wrapping y-axis labels. Default is \code{30}.}

\item{x_end}{Numeric, the endpoint for the x-axis. Default is \code{100}.}

\item{subtitle_wrap}{Numeric, width for wrapping the subtitle text. Default is \code{80}.}
}
\value{
A ggplot object
}
\description{
Generate a bar chart for a single variable from a dataset, with options for
sorting, axis inclusion, and customization.
}
\details{
\code{var} must be submitted as a name, not a character string.
}
\examples{
\dontrun{
# Example usage:
p <- one_bars(data = my_data, var = my_variable)
}

}
