% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bars.R
\name{stacked_bars}
\alias{stacked_bars}
\title{A stacked bar chart}
\usage{
stacked_bars(
  data,
  x_first,
  x_second,
  y,
  sort_by = "first",
  first_fill = "#002856",
  second_fill = "#009AD7",
  bar_width = 0.9,
  y_wrap_width = 20,
  first_label_pos = 20,
  total_nudge = 2.5,
  x_end = NULL,
  x_end_nudge = 1,
  x_mids = NULL
)
}
\description{
Later, allow users to turn totals on and off
}
