% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_priorities_poll}
\alias{theme_priorities_poll}
\title{Generic theme for Priorities Poll charts}
\usage{
theme_priorities_poll(text_size = 12)
}
\description{
Generic theme for Priorities Poll charts
}
